refactor: creaetPages to features/blog
